feat(wip): add support for stateless workflow
