+++ apps/api/src/app/events/dtos/trigger-event-request.dto.ts
@@ -64,7 +64,6 @@ export class TriggerEventRequestDto {
   @IsObject()
   @IsOptional()
   bridge: {
-    workflow: any;
     url: string;
   };
