+++ apps/api/src/app/events/events.controller.ts
@@ -68,9 +68,6 @@ export class EventsController {
     @UserSession() user: UserSessionData,
     @Body() body: TriggerEventRequestDto
   ): Promise<TriggerEventResponseDto> {
-    console.log('HII!!!!');
-    console.log({ body });
-
     const result = await this.parseEventRequest.execute(
       ParseEventRequestMulticastCommand.create({
         userId: user._id,
