+++ apps/worker/package.json
@@ -34,6 +34,7 @@
     "@novu/dal": "workspace:*",
     "@novu/shared": "workspace:*",
     "@novu/stateless": "workspace:*",
+    "@novu/framework": "workspace:*",
     "@novu/testing": "workspace:*",
     "@sentry/node": "^7.40.0",
     "@sentry/tracing": "^7.40.0",
