+++ libs/application-generic/src/utils/require-inject.ts
@@ -6,6 +6,8 @@ import { ChatProviderIdEnum } from '@novu/shared';
 export const requireInject = (inject: RequireInject, moduleRef?: ModuleRef) => {
   if (inject === RequireInjectEnum.RESONATE) {
     return initiateResonateProvider(moduleRef);
+  } else if (inject === RequireInjectEnum.DO_BRIDGE_REQUEST) {
+    return initiateDoBridgeRequestProvider(moduleRef);
   }
 };
 
@@ -39,10 +41,41 @@ const initiateResonateProvider = (moduleRef: ModuleRef) => {
   }
 };
 
+const initiateDoBridgeRequestProvider = (moduleRef: ModuleRef) => {
+  try {
+    if (
+      process.env.NOVU_ENTERPRISE === 'true' ||
+      process.env.CI_EE_TEST === 'true'
+    ) {
+      if (!require('@novu/ee-echo-worker')?.DoBridgeRequest) {
+        throw new PlatformException('Resonate provider is not loaded');
+      }
+
+      return moduleRef.get(require('@novu/ee-echo-worker')?.DoBridgeRequest, {
+        strict: false,
+      });
+    } else {
+      return {
+        execute: () => {
+          return null;
+        },
+      };
+    }
+  } catch (e) {
+    Logger.error(
+      e,
+      `Unexpected error while importing enterprise modules`,
+      'DoBridgeRequest'
+    );
+    throw e;
+  }
+};
+
 type RequireInject = `${RequireInjectEnum}`;
 
 enum RequireInjectEnum {
   RESONATE = 'resonate',
+  DO_BRIDGE_REQUEST = 'do_bridge_request',
 }
 
 export interface IBridgeDigestResponse {
@@ -97,7 +130,7 @@ export type IBridgeStepResponse =
   | IBridgeActionResponse;
 
 export interface IUseCaseInterface<TInput, TResponse> {
-  execute: (arg0: TInput) => TResponse;
+  execute: (arg0: TInput) => Promise<TResponse>;
 }
 
 export interface IUseCaseInterfaceInline {
