+++ libs/dal/src/repositories/notification-template/notification-template.entity.ts
@@ -125,6 +125,8 @@ export class StepVariantEntity implements IStepVariant {
   metadata?: IWorkflowStepMetadata;
 
   shouldStopOnFail?: boolean;
+
+  bridgeUrl?: string;
 }
 
 export class NotificationStepEntity extends StepVariantEntity implements INotificationTemplateStep {
